# Check for duplicate attributes under read lock on insert.